Refactor lLimitNotice, now can receive custom text do display
