+++ src/ui/LimitNotice.tsx
@@ -1,11 +1,11 @@
-import { useTranslation } from "react-i18next";
+interface Props {
+    text: string;
+}
 
-export const LimitNotice = () => {
-  const { t } = useTranslation();
-
+export const LimitNotice = ({text}: Props) => {
   return (
     <div className="w-full bg-yellow-400 text-black py-2 px-4 animate-pulse shadow-md text-center font-semibold">
-      {t("limitReached")}
+      {text}
     </div>
   );
 };
